# services/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tennis Tournament API",
        "version": "0.1"
    },
    "paths": {
        "/api/tournaments": {
            "get": {
                "operationId": "8626d4f0c19b7c5d27dce357833b6b97",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Error retrieving tournaments"
                    }
                }
            }
        },
        "/api/tournaments/{tournament}": {
            "get": {
                "summary": "Get a tournament",
                "operationId": "b44b99a211d160d7c1fa08e4817906d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tournament id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Tournament not found"
                    },
                    "500": {
                        "description": "Error retrieving tournament"
                    }
                }
            }
        },
        "/api/tournaments/date/{date}": {
            "get": {
                "summary": "Get a tournaments by date",
                "operationId": "0cf78352f40b4d2208cc5f1f2f06862d",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "date of start of the tournament",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "date": {
                                "summary": "A date value.",
                                "value": "2021-10-10T00:00:00"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Error retrieving tournament"
                    }
                }
            }
        },
        "/api/tournaments/gender/{gender}": {
            "get": {
                "summary": "Get a tournaments by gender",
                "operationId": "5722b2b1436ea8a91667a7d00d8f15cd",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "path",
                        "description": "gender of the tournament",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A gender value.",
                                "value": "male"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Error retrieving tournament"
                    }
                }
            }
        },
        "/api/tournaments/{tournament}/matches": {
            "get": {
                "summary": "Get a tournaments matches",
                "operationId": "466586aa3f6489d91b841f09e3f3620d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tournament id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Error retrieving matches"
                    }
                }
            }
        },
        "/api/tournaments/{tournament}/winner": {
            "get": {
                "summary": "Get a tournament winner",
                "operationId": "200f68b4f6019d35cfddea633826044e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tournament id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Error retrieving winner"
                    }
                }
            }
        },
        "/api/tournaments/play-tournament": {
            "post": {
                "summary": "Simulate a tournament",
                "operationId": "c05b255d11153ad7bd00be352bf3b4b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "player_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": "[1,2,3,4]"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Error playing tournament"
                    },
                    "500": {
                        "description": "Error playing tournament"
                    }
                }
            }
        }
    }
}